Moved check content/update and ird over to its own command.
